+++ api/features/bootstrap/FeatureContext.php
@@ -1,0 +1,10 @@
+<?php
+
+use Behat\Behat\Context\Context;
+
+/**
+ * FeatureContext
+ */
+class FeatureContext implements Context
+{
+}