+++ api/src/Tenant/Unloader/ServiceUnloader.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace App\Tenant\Unloader;
+
+use App\Entity\Service;
+use Doctrine\ORM\EntityManagerInterface;
+use Ds\Component\Tenant\Entity\Tenant;
+use Ds\Component\Tenant\Loader\Unloader;
+
+/**
+ * Class ServiceUnloader
+ */
+final class ServiceUnloader implements Unloader
+{
+    /**
+     * @var \Doctrine\ORM\EntityManagerInterface
+     */
+    private $entityManager;
+
+    /**
+     * Constructor
+     *
+     * @param \Doctrine\ORM\EntityManagerInterface $entityManager
+     */
+    public function __construct(EntityManagerInterface $entityManager)
+    {
+        $this->entityManager = $entityManager;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function unload(Tenant $tenant)
+    {
+        $builder = $this->entityManager->getRepository(Service::class)->createQueryBuilder('e');
+        $builder
+            ->delete()
+            ->where('e.tenant = :tenant')
+            ->setParameter('tenant', $tenant->getUuid());
+        $query = $builder->getQuery();
+        $query->execute();
+    }
+}