+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## 0.15.0
 
-- Feature [Composer] Upgrade apip dependency to 2.3.4
+- Feature [Composer] Upgrade apip dependency to 2.3.6
+- Feature [Composer] Upgrade symfony dependency to 4.2.0
 - Feature [Fixture] Separate sequence reset fixture from entity fixture
 - Feature [Tests] Add wider range of behat tests
+- Feature [Api] Rename pagination query parameters to `_page` and `_limit`
 
 ## 0.14.0 (2018-10-29)