+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Feature [Fixture] Separate sequence reset fixture from entity fixture
 - Feature [Tests] Add wider range of behat tests
 - Feature [Api] Rename pagination query parameters to `_page` and `_limit`
+- Feature [Config] Config api filter `key` changed from `partial` to `exact` strategy
 
 ## 0.14.0 (2018-10-29)